refactor: 정렬 관련 constants를 sortOptions.ts 파일로 이동
