+++ src/app/collection/page.tsx
@@ -13,7 +13,8 @@ import ShareModal from '@/components/Modal/ShareModal';
 import Pagination from '@/components/Pagination';
 import SortDropdown from '@/components/SortDropdown';
 import { COLLECTION } from '@/constants/API';
-import { ARTWORK_SORT_OPTIONS, ITEMS_PER_PAGE } from '@/constants/pagination';
+import { ITEMS_PER_PAGE } from '@/constants/pagination';
+import { ARTWORK_SORT_OPTIONS } from '@/constants/sortOptions';
 import useDeleteCollection from '@/hooks/serverStateHooks/useDeleteCollection';
 import useGetCollectionArtworks from '@/hooks/serverStateHooks/useGetCollectionArtworks';
 import modalStore from '@/stores/modalStore';
