+++ src/components/ProfilePage/UserArtworkSection/CollectionTab.tsx
@@ -9,10 +9,8 @@ import Icon from '@/components/Icon/Icon';
 import CreateCollectionModal from '@/components/Modal/CreateCollectionModal';
 import Pagination from '@/components/Pagination';
 import SortDropdown from '@/components/SortDropdown';
-import {
-  COLLECTION_SORT_OPTIONS,
-  ITEMS_PER_PAGE,
-} from '@/constants/pagination';
+import { ITEMS_PER_PAGE } from '@/constants/pagination';
+import { COLLECTION_SORT_OPTIONS } from '@/constants/sortOptions';
 import useGetProfileCollectionList from '@/hooks/serverStateHooks/useGetProfileCollectionList';
 import modalStore from '@/stores/modalStore';
 import useToastStore from '@/stores/useToastStore';
