+++ src/components/ProfilePage/UserArtworkSection/LikedTab.tsx
@@ -7,9 +7,9 @@ import OvalButton from '@/components/Button/OvalButton';
 import ArtworkAndCollectionCard from '@/components/Card/ArtworkAndCollectionCard';
 import Pagination from '@/components/Pagination';
 import SortDropdown from '@/components/SortDropdown';
-import { ARTWORK_SORT_OPTIONS, ITEMS_PER_PAGE } from '@/constants/pagination';
+import { ITEMS_PER_PAGE } from '@/constants/pagination';
 import ROUTE from '@/constants/routes';
-import { SORT_OPTIONS } from '@/constants/sortOptions';
+import { ARTWORK_SORT_OPTIONS, SORT_OPTIONS } from '@/constants/sortOptions';
 import useGetLikedArtworkList from '@/hooks/serverStateHooks/useGetLikedArtworkList';
 import useToastStore from '@/stores/useToastStore';
