+++ src/constants/pagination.ts
@@ -1,12 +1,1 @@
 export const ITEMS_PER_PAGE = 12;
-
-export const ARTWORK_SORT_OPTIONS: Record<string, string> = {
-  최신순: 'recent',
-  인기순: 'popular',
-  조회순: 'view',
-};
-
-export const COLLECTION_SORT_OPTIONS: Record<string, string> = {
-  최신순: 'recent',
-  가나다순: 'alphabetic',
-};
