+++ src/constants/sortOptions.ts
@@ -1,1 +1,12 @@
 export const SORT_OPTIONS = ['최신순', '인기순', '조회순'];
+
+export const ARTWORK_SORT_OPTIONS: Record<string, string> = {
+  최신순: 'recent',
+  인기순: 'popular',
+  조회순: 'view',
+};
+
+export const COLLECTION_SORT_OPTIONS: Record<string, string> = {
+  최신순: 'recent',
+  가나다순: 'alphabetic',
+};
